Dev mizera

+++ src/main/java/com/koczy/kurek/mizera/thesisbrowser/service/IThesisService.java
@@ -12,6 +12,6 @@ public interface IThesisService {
 
     List<Thesis> searchTheses(ThesisFilters thesisFilters);
 
-    ThesisDetails getThesisDetails(Long id);
+    ThesisDetails getThesisDetails(int id);
 
 }
